fix(mongo-typesense): rename audit_log to event
