+++ shared/src/typesense/types/event.rs
